requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.5
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
dnspython==2.4.2
feedparser==6.0.10
idna==3.4
kombu==5.3.3
mongomock==4.1.2
mypy==1.7.0
mypy-extensions==1.0.0
packaging==23.2
prompt-toolkit==3.0.40
psycopg2-binary==2.9.9
pymongo==4.6.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sentinels==1.0.0
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.9
